refactor, changing class positioning and deleting CLI
