+++ src/main/java/com/mindlinksoft/recruitment/mychat/ConversationExporter.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat.Tools;
+package com.mindlinksoft.recruitment.mychat;
 
 
 import com.mindlinksoft.recruitment.mychat.Objects.Conversation;
