+++ src/test/java/com/mindlinksoft/recruitment/mychat/ConversationExporterTests.java
@@ -3,7 +3,6 @@ package com.mindlinksoft.recruitment.mychat;
 import com.google.gson.*;
 import com.mindlinksoft.recruitment.mychat.Objects.Conversation;
 import com.mindlinksoft.recruitment.mychat.Objects.Message;
-import com.mindlinksoft.recruitment.mychat.Tools.ConversationExporter;
 import org.junit.Test;
 
 import java.io.FileInputStream;
