--- users.txt
@@ -1,3 +1,0 @@
-senderID: angus generatedID: 44787
-senderID: bob generatedID: 69668
-senderID: mike generatedID: 34131
